refactor : auth와 member 도메인 분리
