+++ src/main/java/com/sparksInTheStep/webBoard/member/presentation/dto/MemberRequest.java
@@ -1,4 +1,4 @@
-package com.sparksInTheStep.webBoard.auth.presentation.dto;
+package com.sparksInTheStep.webBoard.member.presentation.dto;
 
 public record MemberRequest(String nickname, String password) {
 }
